refactor(argus_service.py): Move user-related methods to UserService
