+++ argus/backend/controller/main.py
@@ -6,7 +6,7 @@ from flask import (
 from argus.backend.controller.notifications import bp as notifications_bp
 from argus.backend.service.argus_service import ArgusService
 from argus.backend.models.web import WebFileStorage
-from argus.backend.service.user import login_required
+from argus.backend.service.user import UserService, login_required
 
 LOGGER = logging.getLogger(__name__)
 
@@ -125,7 +125,7 @@ def profile_oauth_github_callback():
         return redirect(url_for("main.error", type=403))
 
     req_code = request.args.get("code", "WTF")
-    service = ArgusService()
+    service = UserService()
     try:
         first_run_info = service.github_callback(req_code)
     except Exception as exc:  # pylint: disable=broad-except
@@ -171,7 +171,7 @@ def upload_file():
         flash(message="No picture provided", category="error")
         return redirect(url_for("main.profile"))
 
-    service = ArgusService()
+    service = UserService()
     filename, filepath = service.save_profile_picture_to_disk(
         picture_name, picture_data, g.user.username)
     service.update_profile_picture(filename, filepath)
@@ -186,7 +186,7 @@ def update_full_name():
     if not new_name:
         flash(message="Incorrect new name", category="error")
     else:
-        service = ArgusService()
+        service = UserService()
         service.update_name(g.user, new_name)
         flash("Successfully changed name!", category="success")
     return redirect(url_for("main.profile"))
@@ -199,7 +199,7 @@ def update_email():
     if not new_email:
         flash("Incorrect new email", category="error")
     else:
-        service = ArgusService()
+        service = UserService()
         service.update_email(g.user, new_email)
         flash("Successfully changed email!", category="success")
     return redirect(url_for("main.profile"))
@@ -222,7 +222,7 @@ def update_password():
         flash("New password doesn't match confirmation!", category="error")
         return redirect(url_for("main.profile"))
 
-    service = ArgusService()
+    service = UserService()
     try:
         service.update_password(
             g.user, old_password=old_password, new_password=new_password)
